Poc swagger codegen maven plugin (#78)
* wip

* update

* update pom

* update

* wip

* wip

* Update

* Modification of package generated by codegen

* wip

* wip

* wip

* Modification of pom xml to remove logback and slf4j

* Wip

* Migration group creation to swagger generation

* Migration get groups

* Migration add users group

* Migration delete group

* Migration get group

* Migration remove user group

* Migration create category

* Migration delete category

* Migration get category

* Update

* update create classification

* update delete classification

* update get classification

* update get classification

* update currency

* update factory

* update third

* update Banking Account

* update Banking transaction

* update Banking transaction

* update balance calculation and currency

* Modification of pmd to remove error processing

* pmd remarks

* update pmd for ci

* update ci

* update ci

* update ci

* pmd remarks

* update ci

* update

* Adding tests

+++ server/src/main/java/fr/finanting/server/model/BankingAccount.java
@@ -6,6 +6,7 @@ import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.Table;
 
+import com.sun.xml.bind.v2.TODO;
 import fr.finanting.server.model.embeddable.Address;
 import fr.finanting.server.model.embeddable.BankDetails;
 import fr.finanting.server.model.mother.MotherGroupUserElement;
@@ -32,7 +33,7 @@ public class BankingAccount extends MotherGroupUserElement {
     private String abbreviation;
 
     @Column(name = "INITIAL_BALANCE", nullable = false)
-    private Integer initialBalance = 0;
+    private Double initialBalance = 0.0;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "DEFAULT_CURRENCY_ID", nullable = false)
